% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{addPseudoCount}
\alias{addPseudoCount}
\title{Add library-size adjusted pseudo-counts}
\usage{
addPseudoCount(data, pseudo.count = 1)
}
\arguments{
\item{data}{A un-normalized count data matrix of shape n x p, where n is the
number of samples and p is the number of features.}

\item{pseudo.count}{A numeric scalar of pseudo-counts to be added to each gene, default: 1.}
}
\value{
counts matrix with adding pseudo counts
}
\description{
Add library-size adjusted pseudo-counts adapted from edgeR
}
