% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/main.R
\name{calcScaleFactor}
\alias{calcScaleFactor}
\title{Generate scale factors based on spike-ins}
\usage{
calcScaleFactor(
  data,
  spike.in.prefix = NULL,
  enrich.group,
  input.id = "Input",
  enrich.id = "Enrich"
)
}
\arguments{
\item{data}{A un-normalized count data matrix of shape n x p, where n is the
number of samples and p is the number of features.}

\item{spike.in.prefix}{A character specify the prefix of spike-in id, e.g.,
"^FB" stands for fly spike-in id, default: NULL.}

\item{enrich.group}{Vector of enrichment group, e.g., c("Input","Enrich","Input","Enrich").}

\item{input.id}{Input library id, must be consistent with \code{enrich.group}, e.g., "Input".}

\item{enrich.id}{Enrich library id, must be consistent with \code{enrich.group}, e.g., "Enrich".}
}
\value{
Vector of scale factors
}
\description{
Generate scale factors based on spike-ins
}
