% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helper.R
\name{synScatter}
\alias{synScatter}
\title{Scatter plot of synthetic RNA ratios}
\usage{
synScatter(ratio.df, syn.meta)
}
\arguments{
\item{ratio.df}{data.frame containing the NCIN ratios of synthetic spike-ins.}

\item{syn.meta}{data.frame of metadata of synthetic spike-ins.}
}
\value{
ggplot
}
\description{
Scatter plot of synthetic RNA ratios
}
